extract arrow result wrapper
